kerncore: improve access check performance
Because the code already requires the region to be ordered by address
and non-overlapping, we can use a binary search here. This makes timing
a lot more consistent -- it's now O(log n) in the region table size
(which is 8 in practice) and then a potential linear scan to the end of
the slice.