kern: let drivers disable double-interrupt behavior
Currently drivers tend to get a spurious interrupt notification just
after handling a real one. This is because the interrupt controller
notices the kernel generic ISR exit while the interrupt condition is
still asserted, because it doesn't expect us to be handling it in task
code. So, it sets the pending bit so the 'new' interrupt condition isn't
forgotten.

This causes an interrupt to occur immediately when we re-enable.

This adds an additional bit for IRQ_CONTROL that drivers can use to
request any pending interrupt status be cleared.

Fixes #536.