arm_m: improve apply_memory_protection performance
This alters the v6/v7 routine to switch off the MPU before updating it,
and switch it back on after, like v8 does. This has allowed me to use
cheaper code to do the update. That, combined with some field ordering
changes to optimize for access order in the generated code, cuts the
cycles spent in this routine by roughly half.