umem: special-case empty copies.
Empty copies are somewhat unusual, but appear in IPC interfaces that
follow certain patterns, or in lease accesses that hit the end.
Previously they were expensive nops. This makes them cheap nops.

This reduces the cost of a zero-byte send by 24% (to about 928 cycles on
a Cortex-M4, if you're curious).