stm32h7-rng: rework driver, add seed error recovery
We have a machine in manufacturing that gets periodic seed errors. The
manual specifies a recovery procedure for seed errors... which we
haven't implemented.

This commit makes us more tolerant of seed errors by attempting
recovery, but not twice in a row. Two seed errors in a row will still
return an error.

This also adds a ringbuf so we can monitor frequency. The ringbuf
increases RAM usage enough that it bumps past the 512-byte boundary, so
I removed all the lines that limit it to 512 (life is too short).