Remove Eq/PartialEq for TaskState.
TaskState is a nontrivial enum. It turns out, we were comparing it often
enough that the compiler thought it'd be good to out-line the comparison
into a callable function.

That function is big and slow because it covers every case. Being
out-lined, it has to be generic and can't use any context to optimize.

So, I've gone through and replaced all TaskState comparisons with
pattern deconstruction. This knocks like 256 bytes off a typical kernel
and speeds up syscall-heavy code by 3% on average.