include the last reason in host-sp-comms reboots
Instead of just setting it to `HostReboot` always, hang onto the last
power off until reaching A0, so that the reason can be sent for the
`set_state` call to reboot, as well as the power off.

Also, just use `StateChangeReason` here instead of our own enum, and add
it to the `host-sp-comms` ringbuf as well.