Disable periodic IPCC zero bytes on Grapefruit (#1988)
In the current IPCC design, the SP sends zero bytes periodically when
the channel is idle to guard against the possibility of losing a frame
terminator. On gimlet this causes the UART FIFO to quickly fill up and
so when the channel wakes up there are around 16 bytes to drain and
things move along. On grapefruit/cosmo, however, there is a 4K FIFO in
the FPGA upstream of the SP here so the natural saturation doesn't (yet)
occur as we'd like. The extra zero bytes also make tracing the bus
harder. Disable them for now.