Move `by_refdes` into `pub(crate)` module (#1994)
This is standalone prep for Cosmo, making the macro useable outside of
the Gimlet inventory.

In addition, we add a variant to the macro which lets the caller control
the array size. This is needed when there are designators of different
lengths in the same inventory slot, e.g. `U123` and `U51`.