transceivers-server: keep LEDs off while not in A0 (#1998)
While Sidecar is not in A0 the transceivers-server will now keep the
LEDs off. The current behavior is that the LEDs only change state when
Dendrite tells them to, and while the Sidecar is not in Dendrite isn't
doing much at all! An A0->A2 transition left the LEDs in their current
state which is confusing for folks who may have eyes on the switch as it
would indicate Dendrite is running when it is in fact not. The system
attention LED is unaffected by this commit.

fixes https://github.com/oxidecomputer/hubris/issues/1995