Standalone spartan-7 loader (#2003)
For Grapefruit and Cosmo, the Spartan-7 FPGA is used as a memory-mapped
peripheral once its bitstream is loaded. It's important that it not die
or be reset, because that will be Very Confusing for any tasks writing
directly into its memory space.

This PR moves loading that FPGA into a standalone task, separate from
`grapefruit-seq` (and the eventual `cosmo-seq`). The new task is
infallible once it enters the Idol runtime loop, and can provide
unforgeable tokens that prove the FPGA was loaded; eventually, I'd like
for FPGA-based peripherals to take those tokens in their constructors.

The code changes in this PR are mostly reorganization and boilerplate,
pulling stuff from `grapefruit-seq` into `drv-spartan7-loader` and
writing `drv-spartan7-loader-api` / Idol files.

Tested on Grapefruit (standalone).