transceivers: pull in power enable status fix (#2009)
This pulls in an update QSFP front IO FPGA at
https://github.com/oxidecomputer/quartz/commit/9209c858a8c91d7ab4d81e146803dcc4f4b8c280.
This FPGA change makes it more clear if power is actually enabled to a
module or not by splitting out the software power enable register from a
hardware readback power enable register. In practice the FPGA gated on
the enable bit being set **and** module presence, but that was not
reflected in the `power_enable` field of module status. The only
adjustment needed in the server is that we must write to the new
`QSFP_SW_POWER_ENx` registers instead of the old `QSFP_POWER_ENx`
registers which are now read only.