Peripheral codegen for Grapefruit / Cosmo (#2010)
This PR adds a new target to `build-fpga-regmap`, generating Rust code
for a memory-mapped peripheral. For Grapefruit and Cosmo, the Spartan-7
FPGA is connected over FMC, so we can treat peripherals implemented in
the FPGA as native(ish). This PR uses the new codegen paths for
`cosmo-hf` and `grapefruit-seq`, replacing raw addresses. The peripheral
types take a `drv_spartan7_loader_api::Spartan7Token`, which proves that
the FPGA is loaded and running.

Other changes that came along for the ride:

- Grapefruit FPGA files (bitstream and JSON) are moved to
`drv/spartan7-loader/grapefruit/` subfolder. I also updated to a more
recent build (latest `main`), and removed the HTML file (since the JSON
is smaller + has the same information).
- `syn` is bumped to version 2, which requires a few miscellaneous
changes