MMIO-attached peripherals (#2024)
This lets us specify a set of FMC-attached **virtual peripherals**,
which are implemented in an FPGA.

A manifest may specify an `[mmio]` configuration object, with a base
address and register map (which is a JSON file). If present, `xtask`
parses the JSON file and adds its peripherals to the `chip.toml`'s
peripheral list, prefixed with `mmio_`. Tasks can include these
peripherals in their `uses = [...]` list.

Instead of copying `chip.toml` into the output archive, we re-serialize
the `peripherals` map, so that it includes these virtual peripherals.
This means that `humility map` works fine:

```
DESC       LOW          HIGH          SIZE ATTR   ID TASK
...elided...
0x08005820 0x60000000 - 0x600000ff     256 rw-d-- 24 [mmio_base] fmc_demo, hf
0x0800585c 0x60000100 - 0x600001ff     256 rw-d-- 24 [mmio_spi_nor] fmc_demo, hf
0x08005834 0x60000200 - 0x600002ff     256 rw-d-- 24 [mmio_espi] fmc_demo
0x08005848 0x60000300 - 0x600003ff     256 rw-d-- 12 [mmio_sgpio] grapefruit_seq, fmc_demo
```

(it does lose the comments in the original `chip.toml`; raise an
objection if you think that's a big deal)

`drv-stm32h7-fmc-demo-server` is now configured to use each of the FPGA's
peripherals, instead of the large `fmc_nor_psram_bank_1` region, because
Humility doesn't play nice with multiple peripherals starting at the same
address. If we are still using this demo task with > 8 peripherals, then we can
reconsider.