Don't modify active slot while updating host flash (#2043)
We don't want programming the host flash to _also_ set the slot used at
boot (see #1958 for details), because interruptions could leave us in an
unfortunate state.

This PR adds new `page_program_dev` and `sector_erase_dev` functions,
which operate on a particular slot (passed as an argument), but are
guaranteed to leave the active slot unchanged when they exit. In other
words, they switch to a particular slot, perform an action, then restore
the previous slot before returning the result.

In `control-plane-agent`, we now use those functions during host flash
updates (previously, we set the host flash slot during
`HostFlashUpdate::prepare` then assumed it wouldn't change).

There's one behavioral change: at the end of the update, we have not
(implicitly) switched to the new slot. Looking at Omicron, it looks like
we don't rely on this behavior, but I'd appreciate a second opinion.

Tested on `niles` and it works as expected (I see `HostFlash.get_dev`
stay the same during the update).