minibar: automatically setup the PCIe refclk buffer (#2048)
Most of the noise here is me adjusting the case of some register fields.
The real change here is to some of the reset values of fields in the
PCIE_REFCLK_CTRL register in the FPGA to get the refclk from the sled to
Minibar's slot. On startup the SP clears the "power down" line to the
clock buffer, at which point it samples its strappings and begins
operation.