"humility qspi -D" is sluggish on...everything else (#2053)
_No description provided._

So, `humility qspi` and auxflash commands are Really Slow on Cosmo. It
turns out that this is because Hiffy has a magical list of boards that
are allowed to go fast, and we never put Cosmo on it. Boards on the
go-faster list are allowed to use a much larger buffer for data
transferred through Hiffy than boards not on the go faster list.

Controlling this based on `target_board` seems quite unfortunate.
Subsequent revisions of the same board need to be added to the list, and
it's easy to forget to do that when there's a new board revision and
regress hiffy performance again (c.f. #1498). Instead, let's control
this via a Cargo feature, which can at least be set in the `base.toml`
for a board and inherited by every subsequent revision of the board
without having to remember to do that.

I had wanted to just always enable go-faster mode on STM32H7 but I let
@cbiffle talk me out of it.

---------

Co-authored-by: Aapo Alasuutari <aapo.alasuutari@gmail.com>