Clear LM5066 faults on `power` startup (#2073)
The LM5066 has strong opinions on what error flags should be set on
startup:

- `STATUS_BYTE` says the VIN fault flag is set by default, no comment
  about autoclearing; flags should be cleared with `CLEAR_FAULTS`
- `STATUS_WORD` says VIN fault flag is set by default, and is cleared
  when input voltage > UVLO threshold
- `STATUS_INPUT` says the VIN _warning_ flag is set by default (the VIN
  fault flag is 0 by default), and is cleared when input voltage > UVLO
  threshold

In practice, we see the fault flag set in all three of these registers,
as well as `!PowerGood`.

This PR adds a `CLEAR_FAULTS` call to the `power` task startup, to
persuade the power-good bit to actually reflect reality. We don't expect
the `power` task to restart, so this should be a one-off occurrence (and
genuine faults later on will not be affected).