Add Cosmo SPD driver (#2094)
The SPD driver on Cosmo doesn't need to be an I2C target, but we still
want to read data from the SPDs into Packrat – both to prove that it
works, and (eventually) to report them to the host as inventory devices.
We also need SPD communication to read temperatures for the thermal
loop.

To make life harder, SPD size changed between DDR4 and DDR5 (512 → 1024
bytes). This PR modifies our existing Packrat implementation to make it
generic. There's a new `struct SpdData<const NUM_DIMMS: usize, const
DATA_SIZE: usize>`, which is used for both Gimlet and Cosmo (with
different values for `DATA_SIZE`).

---------

Co-authored-by: Nathanael Huffman <nathanael@oxidecomputer.com>