Stop mentioing `A1` state outside the sequencer
`A1` is supposed to be a transitory state from `A2` onto `A0`.
Issue #1267 flagged that we were never broadcasting this information
but also made the (modest) suggestion that we could also just
remove it.

It turns out, nobody is attempting to do anything significant when
the system is in A1. Remove this from the public API to make everything
a little easier on the clients. The sequencer still tracks A1
internally.