only allow pcie backplane clock when in lab configurations (#2113)
To support minibar-attached PCIe NIC we need to have a clock running out
the backplane, but we don't want to do this in dev or lab images which
we may run in EMC testing scenarios. The sidecar uses a clock-less PCIe
config (SRIS) and we don't want or need to send the clock out the
backplane in the normal use cases.

If one needed to run a dev image and still wanted the clock output, the
register can be turned on in this case manually.

-lab images will do this by default which will support most minibar
configurations and manufacturing stations.

---------

Co-authored-by: Matt Keeter <matt@oxide.computer>