Introduce `compute-sleds` release target
With the addition of `cosmo`, we've switched to the term `compute-sled`
to describe gimlet and cosmo together. Add this as a release target.
We can keep the old `gimlet` target for backwards compatibility.
If we end up really needing a `cosmo` only release we can add that
as a target too.