stm32xx-i2c-server: stop regenerating SYS handle
This switches the code to pass the handle to Sys around, instead of
using syscalls to re-derive it whole-cloth on each use. This is now
closer to the behavior of our other drivers, making it somewhat cheaper
at runtime -- but more importantly, making it a lot more obvious which
routines can have IPC implications.