stm32xx-i2c: reset controller on startup.
The existing code enabled clock to the I2C peripheral and cleared its
reset condition. This is all well and good if we've just booted, but if
the driver has _crashed,_ it's finding the I2C peripheral in an
arbitrary state -- potentially with a transaction in progress. In this
case, the enable sequence it was using was a no-op.

This is problematic if your next action is to take over the GPIO pins
and wiggle them, something the I2C controller _really_ does not like
(when it's enabled).

This commit resets the I2C controllers into a known state when starting
the driver.