stm32xx-i2c: break up reset operation
The code formerly had a single concept of a "reset." This seemed
confusing on several levels:

- It aliases the `BusReset` error but has nothing to do with it.

- It also aliases the system "reset the peripheral" concept, but does
  _not_ reset the peripheral.

This concept needed to break in two in order to wrap the SCL wiggle
code, to fix #1821. While breaking it, I started changing the names to
better reflect the actual function of the routines: they halt the I2C
protocol engine, without resetting the overall peripheral.

Fixes #1821 (probably)