always reserve space for loss
Currently, `snitc-core` will panic on reads if we have lost data and
there is no space for an additional loss record in the buffer. This is
because reads while the store is in the `Losing` state will attempt to
insert a loss record to report that data has been lost. However, if
there is no space in the queue, inserting the loss record will panic.

This commit fixes this by reserving space for one additional loss
record at all times. We will no longer admit new messages into the
buffer if they mean we cannot encode a loss record when more messages
come in. This, unfortunately, means that we can no longer admit entries
that fill the entire queue, due to this reservation.