shrink ringbuf a bit
previously, we had a ringbuf entry that contained both the current and
requested restart IDs, and was therefore 256B --- meaning the ringbuf
was an array of 16 entries that were at least 257B each (including the
enum discriminant). that's like 4KiB of ringbuf, which is quite
uncomfortable. by moving things around a bit so that we have separate
entries for "we got a request with this restart ID" and "restart IDs
didn't match", we can make the max sized ringbuf entry 128B (plus an
enum discriminant), which basically cuts the size of the ringbuf in
half.