Remove outdated references from README
There are currently a few things in the README which are incorrect or no
longer relevant:

- There is no longer a `stage0/` directory in the repository (I presume
this has been subsumed by `bootleby`)
- The `chip/` directory is actually called `chips/`
- Various references to OpenOCD/pyOCD being required development
dependencies. These are no longer necessary, as `humility` now uses
`probe-rs` instead.

I didn't touch a couple bits of sample output from `pyocd` in the
LPCXpresso55S69 documentation, as I don't have that board on hand and
couldn't generate the equivalent examples for `probe-rs`' commands. If
someone has that board and wants to take the time to produce the
equivalent example output, I'm happy to add them here.