xtask: Override default caboose version from env
When flashing an engineering Hubris image on an SP that will participate
in a control-plane driven software update, it is necessary to populate
the `VERS` key in the caboose must be present so that the update
machienry doesn't get confused. If it is desirable for the engineering
image to not be overwritten during an update, the verson must also be
set to the version expected by the system. Being able to do this is
useful for testing Hubris changes in integration with the rest of the
system.

Historically, such images have been created by hacking up the `xtask
dist` code to write the desired version string when creating the default
caboose. I thought that this seemed like a common enough thing to want
to do that there should just be a way to ask for it without having to
change the build system. Therefore, this branch adds a
`HUBRIS_CABOOSE_VERS` env var that will override the `VERS` key in the
caboose if building an image that has a default caboose.